moved test part to seperate .py file - moved further functions to _TMC_2209_move.py
