+++ src/TMC_2209/TMC_2209_StepperDriver.py
@@ -42,11 +42,16 @@ class TMC_2209:
     )
 
     from ._TMC_2209_move import (
-        set_max_speed, set_max_speed_fullstep, get_max_speed, set_acceleration,
-        set_acceleration_fullstep, get_acceleration, stop, get_movement_phase,
-        run_to_position_steps, run_to_position_revolutions, run_to_position_steps_threaded,
-        run_to_position_revolutions_threaded, wait_for_movement_finished_threaded,
-        run, distance_to_go, compute_new_speed, run_speed, make_a_step
+        set_movement_abs_rel, get_current_position, set_current_position, set_max_speed,
+        set_max_speed_fullstep, get_max_speed, set_acceleration, set_acceleration_fullstep,
+        get_acceleration, stop, get_movement_phase, run_to_position_steps,
+        run_to_position_revolutions, run_to_position_steps_threaded,
+        run_to_position_revolutions_threaded, wait_for_movement_finished_threaded, run,
+        distance_to_go, compute_new_speed, run_speed, make_a_step
+    )
+
+    from ._TMC_2209_test import (
+        test_dir_step_en, test_step, test_uart, test_stallguard_threshold
     )
 
     tmc_uart = None
@@ -196,18 +201,6 @@ class TMC_2209:
 
 
 
-    def set_movement_abs_rel(self, movement_abs_rel):
-        """
-        set whether the movment should be relative or absolute by default.
-        See the Enum MovementAbsoluteRelative
-
-            Paramters:
-                movement_abs_rel (enum): whether the movment should be relative or absolute
-        """
-        self._movement_abs_rel = movement_abs_rel
-
-
-
     def set_motor_enabled(self, en):
         """
         enables or disables the motor current output
@@ -325,28 +318,6 @@ class TMC_2209:
             self.tmc_logger.log(str(sg_results),Loglevel.DEBUG)
 
         self.tmc_logger.log("---", Loglevel.INFO)
-
-
-
-    def get_current_position(self):
-        """
-        returns the current motor position in microsteps
-
-            Returns:
-                current_pos (bool): current motor position
-        """
-        return self._current_pos
-
-
-
-    def set_current_position(self, new_pos):
-        """
-        overwrites the current motor position in microsteps
-
-            Parameters:
-                new_pos (bool): new position
-        """
-        self._current_pos = new_pos
 
 
 
@@ -558,181 +529,3 @@ class TMC_2209:
         if time.time()<=self._starttime+self._sg_delay and self._sg_delay != 0:
             return
         self._sg_callback()
-
-
-
-    def test_dir_step_en(self):
-        """
-        tests the EN, DIR and STEP pin
-        this sets the EN, DIR and STEP pin to HIGH, LOW and HIGH
-        and checks the IOIN Register of the TMC meanwhile
-        """
-        pin_dir_ok = pin_step_ok = pin_en_ok = True
-
-        GPIO.output(self._pin_step, GPIO.HIGH)
-        GPIO.output(self._pin_dir, GPIO.HIGH)
-        GPIO.output(self._pin_en, GPIO.HIGH)
-        time.sleep(0.1)
-        ioin = self.read_ioin()
-        if not ioin & tmc_reg.io_dir:
-            pin_dir_ok = False
-        if not ioin & tmc_reg.io_step:
-            pin_step_ok = False
-        if not ioin & tmc_reg.io_enn:
-            pin_en_ok = False
-
-        GPIO.output(self._pin_step, GPIO.LOW)
-        GPIO.output(self._pin_dir, GPIO.LOW)
-        GPIO.output(self._pin_en, GPIO.LOW)
-        time.sleep(0.1)
-        ioin = self.read_ioin()
-        if ioin & tmc_reg.io_dir:
-            pin_dir_ok = False
-        if ioin & tmc_reg.io_step:
-            pin_step_ok = False
-        if ioin & tmc_reg.io_enn:
-            pin_en_ok = False
-
-        GPIO.output(self._pin_step, GPIO.HIGH)
-        GPIO.output(self._pin_dir, GPIO.HIGH)
-        GPIO.output(self._pin_en, GPIO.HIGH)
-        time.sleep(0.1)
-        ioin = self.read_ioin()
-        if not ioin & tmc_reg.io_dir:
-            pin_dir_ok = False
-        if not ioin & tmc_reg.io_step:
-            pin_step_ok = False
-        if not ioin & tmc_reg.io_enn:
-            pin_en_ok = False
-
-        self.set_motor_enabled(False)
-
-        self.tmc_logger.log("---")
-        if pin_dir_ok:
-            self.tmc_logger.log("Pin DIR: \tOK")
-        else:
-            self.tmc_logger.log("Pin DIR: \tnot OK")
-        if pin_step_ok:
-            self.tmc_logger.log("Pin STEP: \tOK")
-        else:
-            self.tmc_logger.log("Pin STEP: \tnot OK")
-        if pin_en_ok:
-            self.tmc_logger.log("Pin EN: \tOK")
-        else:
-            self.tmc_logger.log("Pin EN: \tnot OK")
-        self.tmc_logger.log("---")
-
-
-
-    def test_step(self):
-        """
-        test method
-        """
-        self.set_direction_pin(1)
-
-        for _ in range(100):
-            self._current_pos += 1
-            GPIO.output(self._pin_step, GPIO.HIGH)
-            time.sleep(0.001)
-            GPIO.output(self._pin_step, GPIO.LOW)
-            time.sleep(0.01)
-
-
-
-    def test_uart(self):
-        """
-        test method
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("TEST UART")
-        result = self.tmc_uart.test_uart(tmc_reg.IOIN)
-
-        snd = result[0]
-        rtn = result[1]
-
-        self.tmc_logger.log("length snd: "+str(len(snd)), Loglevel.DEBUG)
-        self.tmc_logger.log("length rtn: "+str(len(rtn)), Loglevel.DEBUG)
-
-        if len(rtn)==12:
-            self.tmc_logger.log("""the Raspberry Pi received the sended
-                                bits and the answer from the TMC""",Loglevel.INFO)
-        elif len(rtn)==4:
-            self.tmc_logger.log("the Raspberry Pi received only the sended bits",
-                                Loglevel.INFO)
-        elif len(rtn)==0:
-            self.tmc_logger.log("the Raspberry Pi did not receive anything",
-                                Loglevel.INFO)
-        else:
-            self.tmc_logger.log("the Raspberry Pi received an unexpected amount of bits: "+
-                     str(len(rtn)), Loglevel.INFO)
-
-        if snd[0:4] == rtn[0:4]:
-            self.tmc_logger.log("""the Raspberry Pi received exactly the bits it has send.
-                     the first 4 bits are the same""", Loglevel.INFO)
-        else:
-            self.tmc_logger.log("""the Raspberry Pi did not received the bits it has send.
-                     the first 4 bits are different""", Loglevel.INFO)
-
-
-        self.tmc_logger.log("complete", Loglevel.DEBUG)
-        self.tmc_logger.log(str(snd.hex()), Loglevel.DEBUG)
-        self.tmc_logger.log(str(rtn.hex()), Loglevel.DEBUG)
-
-        self.tmc_logger.log("just the first 4 bits", Loglevel.DEBUG)
-        self.tmc_logger.log(str(snd[0:4].hex()), Loglevel.DEBUG)
-        self.tmc_logger.log(str(rtn[0:4].hex()), Loglevel.DEBUG)
-
-
-        self.tmc_logger.log("---")
-        return True
-
-
-
-    def test_stallguard_threshold(self, steps):
-        """
-        test method for tuning stallguard threshold
-        run this function with your motor settings and your motor load
-        the function will determine the minimum stallguard results for each movement phase
-
-            Parameters:
-                steps (int): amount of steps; can be negative
-        """
-
-        self.tmc_logger.log("---", Loglevel.INFO)
-        self.tmc_logger.log("test_stallguard_threshold", Loglevel.INFO)
-
-        self.set_spreadcycle(0)
-
-        min_stallguard_result_accel = 511
-        min_stallguard_result_maxspeed = 511
-        min_stallguard_result_decel = 511
-
-        self.run_to_position_steps_threaded(steps, MovementAbsRel.RELATIVE)
-
-
-        while self._movement_phase != MovementPhase.STANDSTILL:
-            stallguard_result = self.get_stallguard_result()
-
-            self.tmc_logger.log(str(self._movement_phase) + " | " + str(stallguard_result),
-                     Loglevel.INFO)
-
-            if (self._movement_phase == MovementPhase.ACCELERATING and
-               stallguard_result < min_stallguard_result_accel):
-                min_stallguard_result_accel = stallguard_result
-            if (self._movement_phase == MovementPhase.MAXSPEED and
-               stallguard_result < min_stallguard_result_maxspeed):
-                min_stallguard_result_maxspeed = stallguard_result
-            if (self._movement_phase == MovementPhase.DECELERATING and
-               stallguard_result < min_stallguard_result_decel):
-                min_stallguard_result_decel = stallguard_result
-
-        self.wait_for_movement_finished_threaded()
-
-        self.tmc_logger.log("---", Loglevel.INFO)
-        self.tmc_logger.log("min StallGuard result during acceleration: " +
-                 str(min_stallguard_result_accel), Loglevel.INFO)
-        self.tmc_logger.log("min StallGuard result during maxspeed: " +
-                 str(min_stallguard_result_maxspeed), Loglevel.INFO)
-        self.tmc_logger.log("min StallGuard result during deceleration: " +
-                 str(min_stallguard_result_decel), Loglevel.INFO)
-        self.tmc_logger.log("---", Loglevel.INFO)
